+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Read: ~100 records/ms
 - Add a "revisions" mode to history.  Now history records can be set per row instead of database wide.
 - Add `revisionComment` function to standard queries allowing you to add information about each revision into the datastore.
 
+## [0.8.8] 6-3-2017
+- Made `manualExec` a public method of `query`, allowing you to pass the query object itself instead of using `.where().join()...`.
+- Added `*` event to trigger on all database events.
+- Made selected table variable public.
+- Fixed event handler always showing the event type as "error".
+
 ## [0.8.72] 6-3-2017
 - BREAKING CHANGE: `describe` queries have been reshaped and now also return the primary key and number of rows of the selected table.
 - Fixed an issue with ORM updates on deletes.